requirements.txt
flask
pyboy
openai
requests
keyboard
pillow
websockets == 10.4
click
pyyaml